align exmples includes
