+++ examples/C/color/rs-color.c
@@ -9,7 +9,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdio.h>
-#include "../example.h"
+#include "example.h"
 
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
