+++ examples/C/depth/CMakeLists.txt
@@ -9,6 +9,7 @@ project(RealsenseExamples-Depth)
 set(CMAKE_EXPORT_COMPILE_COMMANDS 1)
 
 add_executable(rs-depth rs-depth.c ../example.h)
+include_directories(../../C)
 target_link_libraries(rs-depth ${DEPENDENCIES})
 set_target_properties (rs-depth PROPERTIES
 FOLDER "Examples/C"
