+++ examples/align-advanced/rs-align-advanced.cpp
@@ -2,9 +2,7 @@
 // Copyright(c) 2017 Intel Corporation. All Rights Reserved.
 
 #include <librealsense2/rs.hpp>
-#include "../example.hpp"
-#include <imgui.h>
-#include "imgui_impl_glfw.h"
+#include "example-imgui.hpp"
 
 #include <sstream>
 #include <iostream>
