+++ examples/capture/CMakeLists.txt
@@ -9,7 +9,7 @@ if(BUILD_GRAPHICAL_EXAMPLES)
     add_executable(rs-capture rs-capture.cpp ../example.hpp)
     set_property(TARGET rs-capture PROPERTY CXX_STANDARD 11)
     target_link_libraries(rs-capture ${DEPENDENCIES})
-    include_directories(../ ../../third-party/tclap/include)
+    include_directories(../ ../../third-party/tclap/include ../../examples)
     set_target_properties (rs-capture PROPERTIES FOLDER "Examples")
 
     install(TARGETS rs-capture RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
